Composer: improve CS requirements
PHP_CodeSniffer is a dependency of the external WPCS, PHPCompatibility(WP) and VIPWPCS standards.

There is no need to explicitly require it for this repo and not doing so reduces the risk of version constraint conflicts with one of the external standards.

Also, makes sure that the PHPCompatibilityWP standard is pulled in at a version within the latest major.