QA: use strict type checking when using `in_array()`
Strict type comparisons should be used as a rule, with loose type comparisons being the exception.

This is especially important when comparing variables which are expected to be strings, as when one of the two operands - for whatever reason, other plugins interfering, errors etc - is not a string, the other operand will be juggled to the type of the first operand which can lead to unexpected results and hard to debug bugs.

For array functions which do loose type comparisons, setting the third `$strict` parameter to `false` can be seen as a clear indication that this is a conscious, well thought out decision by the developer.
In all other cases, `$strict` `true` should be used.

Ref: http://php.net/manual/en/function.in-array.php