Travis: only lint against high/low PHP and run tests on more PHP versions
As it was, the unit tests were only being run against PHP 7.4, while the plugin is supposed to be compatible with PHP 5.6 - 8.0.
To safeguard that properly, the tests should be run against all the supported PHP versions.

At this time, however, it is not (yet) possible to run the tests against PHP 8 due to the PHPUnit version being used (PHPUnit 5-7) in combination with certain tests features.
This will be fixed in the near future in a separate PR.

So, for now, this commit makes the following changes to the CI setup:
* Run the tests against all currently supported PHP versions, with the exception of PHP 8.0 which will be enabled at a later point in time.
* Only lint the code against high/low PHP versions, including against "future" PHP versions, like PHP 8.1 (nightly).
    As the tests are being run on all PHP versions, we don't really need to lint against all PHP versions anymore as most parse errors would be discovered by the tests anyhow. So linting against the highest and lowest supported PHP versions, including against PHP version for which the test runs are not yet enabled will be sufficient.