Composer/Travis: remove sniffing against VIPCS
As discussed.

The VIPCS rulesets have a couple of drawbacks:
1. As it stands, they can not be combined with WordPressCS as they will silence various sniffs from WPCS which we _do_ want to receive notice from.
2. The sniffs themselves are not always that accurate.
3. The sniffs will, at times, recommend using functionality only available on the VIP platform, which means that certain "issues" are not fixable.

It is the intention to include select sniffs - not the rulesets - from VIPCS is a future version of YoastCS.

With that in mind, the sniffing against VIPCS is being removed, in favour of inheriting sniffing against select VIP sniffs via YoastCS in the future.