CS: use brackets around condition
This is a little bit awkward as in most cases in this changeset, there aren't explicit conditions (which is a bad thing).

Generally speaking, ternaries should be used only sparingly and if used, should use explicit conditions and those conditions should be surrounded by parentheses to prevent potential issues with operator precedence.

Even so, even when not using explicit conditions, the sniff I'm addressing with this PR enforces parentheses around the condition part of the ternary.

Where it wouldn't break the code, I've changed the code to not use a ternary.
Where it was obvious what the condition was which was being tested, I've made the condition explicit.

In all other cases, I've - for now - just added parentheses around the "condition" part of the ternary to get it to pass the CS check.

I'd strongly recommend that each of these cases be reviewed individually with an aim at refactoring to better code.