Remove composer.lock file
As per our discussion.

This project does not have run-time dependencies (`require`), only development dependencies (`require-dev`).

When a project has run-time dependencies, and especially when it is subsequently distributed in a manner without dependency management (i.e. zip file plugin distribution), it can be very important to make sure that the versions used of the run-time dependencies are locked to specific versions with which the plugin has been tested thoroughly.

However, that is not the case for this plugin, which only has dev dependencies. In that case, having a `composer.lock` can make things like running the tests on multiple different PHP version which require different versions of the underlying test tools **_more difficult_**.

With that in mind, I'm removing the `composer.lock` file and adding it to the `.gitignore` file to prevent it from being committed again.

If at some point in the future, any of the dev dependencies would need to be "locked" at a specific version (or range), that can be achieved by making the version constraint in the `composer.json` file stricter, i.e. using something like `2.3.*` as a constraint instead of `^2.3`.

If at some point in the future, a non-dev, run-time dependency would be introduced, the decision to remove the `composer.lock` file can be revisited.