Tests: switch over to the Yoast\WPTestUtils\BrainMonkey\YoastTestCase
All test classes for the BrainMonkey based unit tests in the Duplicate Post test suite have a Duplicate Post native `TestCase` as a parent class.

This switches the parent of that `TestCase` from the PHPUnit native TestCase to the `Yoast\WPTestUtils\BrainMonkey\YoastTestCase`.

The `YoastTestCase`:
* Inherits PHPUnit cross-version compatibility from the PHPUnit Polyfills TestCase.
* Inherits the BrainMonkey set up and teardown and Mockery expectation counting as assertions from the BrainMonkey TestCase.
* And natively adds a number of stubs for commonly used WP functions.

This switch over includes:
* Removing the addition of function stubs which are already added via the `YoastTestCase` class.
* Removing the `tearDown()` method. This will be inherited from the `YoastTestCase` now.
* Renaming the `setUp()` method to `set_up()` in various test classes for PHPUnit cross-version compatibility.
* Makes the `set_up()` methods `protected` to match the visibility in PHPUnit itself.

There is one "exception", which is the `Post_Duplicator_Test` class.
Various tests in that class set expectations for the WP `sanitize_text_field()` function, which is stubbed in the `YoastTestCase` and as a function can't be both stubbed as well as have expectations set for it, this would lead to failing tests.
So for that particular class, the parent class is switched over to the more generic ``Yoast\WPTestUtils\BrainMonkey\TestCase` class, which has the same features as the `YoastTestCase`, with the exception of the additional stubs.