TestCase: remove the generic stubbing of the translation and escaping functions
Only a limited number of tests - 66 out of 282 - need these stubs, so we may as well only load them if and when needed.

BrainMonkey offers a `stubTranslationFunctions()` function which stubs all nearly translation functions and includes escaping for the "translate and escape" functions.
Similarly, it contains a `stubEscapeFunctions()` function.

The WP Test Utils base class for BrainMonkey offers an alternative implementation for both of these, which doesn't do any escaping, but will always return the original value unchanged.

For each of the 66 tests which actually need either the translation functions or the escaping functions, or both, a call to one or both of the wholesale stubbing functions has been put in place.

As a rule of thumb, these function calls have been added to a test `set_up()` method if more than half the tests in the class need these functions.

Otherwise, the function call(s) have been added to the individual test.