PHPCS: remove use of deprecated WPCS native ignore annotations
The use of WPCS native ignore annotations like `// Input var okay` has been deprecated in WordPressCS 2.0.0 and support for these annotations will be removed in WPCS 3.0.0.

With that in mind, there annotations should be removed and if necessary replaced by the PHPCS native ignore annotations, i.e. `// phpcs:ignore....`.

In the case of the `// Input var okay` annotations I'm removing in this commit, none of them need replacing as none are actually ignoring an error/warning.

These annotations may have entered the codebase via copy/paste or may have ignored violation reports in the past, but WPCS does not actually throw violations for these lines as they are now, anymore.

Refs:
* https://github.com/WordPress/WordPress-Coding-Standards/releases/tag/2.0.0-RC1
* https://github.com/WordPress/WordPress-Coding-Standards/wiki/Whitelisting-code-which-flags-errors