CS: consistency for file includes
The `DUPLICATE_POST_PATH` constant is declared early in the plugin bootstrap file.
For consistency, it is recommended to use this constant throughout the application for file includes.

Note: the `DUPLICATE_POST_PATH` constant is based on `plugin_dir_path()` which returns the path with a trailing slash, so leading slashes in the "file name" part of the paths have been removed.