Tests: fix broken setup of mocked WP_Role objects
This fixes three different issues with the mocking of the WP_Role objects.

1. Incorrect implementation of `allows()`.

As per the Mockery documentation about the `allows()` syntax, the use of `allows()` can take two forms:
1. Passing it an array with `method_name` => `return_value` entries.
2. Not passing it any arguments, but chaining `method_name` and `andReturn( 'return value' )` to it.

In the Duplicate Post `TestCase` class method [1] was being used, but implemented incorrectly.

The `return_value` should be an actual value, not a callback which will be called to get at the value.

In practice, that meant that the mocked `WP_Role` `hap_cap()`, `add_cap()` and `remove_cap()` methods were returning a closure (object instance of the `Closure` class and not the expected return value.
The only reason this hadn't led to failing tests before is the prevalence of implicit loose type comparisons in the code base, i.e. `if ( $something ) {}`.

Ref: http://docs.mockery.io/en/latest/reference/alternative_should_receive_syntax.html#allows

2. Void method(s) returning a value.

The WP `WP_Role::add_cap()` method is, just like the `WP_Role::remove_cap()` method, a `void` method, i.e. it does not return a value.

As it was, the `add_cap()` was returning a closure, but based on the return value of the closure, it was intended for the method to return `true`.

Changing this to `null` now in an attempt to more closely match the original method.

Ref: https://developer.wordpress.org/reference/classes/wp_role/add_cap/

3. Incorrect use of `makePartial()`.

As per the documentation:

> Partial doubles are useful when we want to stub out, set expectations for, or spy on some methods of a class, but run the actual code for other methods.

As the actual `WP_Role` class is not available to the test suite as WP will not be loaded, making these mocks "partials" is useless.

Ref: http://docs.mockery.io/en/latest/reference/creating_test_doubles.html#partial-test-doubles