TestCase: move WP_Role mocking/stubbing to separate function
The mocking of the WP_Role objects and the stubbing of the `get_role()` function is only used by a few tests, so instead of of setting these up for every single test being run, only set these up selectively for those tests which actually use this functionality.

Includes removing the `$roles` property which is unused.