Tests: fix bug in Options_Form_Generator_Test::test_generate_roles_permission_list()
As per the mock of the WP_Role objects in the `Yoast\WP\Duplicate_Post\Tests\TestCase::setUp()` method, the `has_cap()` method will respectively return `true` for the `Editor` role and `false` for the `Administrator` and the `Subscriber` role.

Based on that, the expected output for the `Options_Form_Generator_Test::test_generate_roles_permission_list()` test should have only the `Editor` checkbox checked.

This fixes the incorrectly checked checkbox for the `Administrator` role.