PHPCS: remove inline ignore annotations for nonce verification
These issues should not be ignored, but fixed. A throughout review of the codebase is needed to verify that nonces are used everywhere and verified correctly.

To that end:
* A "Technical Debt" issue has been (or will be) created by Enrico to add this to the planning for the team.
* A temporary exclusion for the existing issues which were being flagged by PHPCS has been added to the ruleset.

When the review takes place, the temporary exclusion in the PHPCS ruleset can be commented out to see all the issues.
For those issues where nonce verification is handled correctly, the inline ignore annotation can be brought back, but should contain a reference to where the nonce verification is done. Example:

```php
// phpcs:ignore WordPress.Security.NonceVerification -- Reason: nonce is verified earlier in the code flow in function ClassName::noncecheck().
$id   = \intval( \wp_unslash( $_GET['post'] ) );
```