Tests: bug fix - add missing function call mocking
This is a result of an audit to check whether all tests which test functions which use any of the static methods in the `Utils` class, correctly mock the function from the `Utils` class.

In most cases, this was done correctly. However, the audit turned up these two tests where the function calls to the `Utils` class within the function under test were not being mocked.