Options_Form_Generator::sort_taxonomy_objects(): improve sort order and fix PHP 8 compatibility
The `u*sort()` functions in PHP have always been expected to return an integer: a negative integer for when A should be considered "less than" B, a `0` when they should be considered as equal and positive integer for when B should be considered "more than" A.

The `Options_Form_Generator::sort_taxonomy_objects()` did not comply with this and returned a boolean value instead.

As of PHP 8.0, PHP checks more strictly on the callback implementation being correct and will throw a `usort(): Returning bool from comparison function is deprecated, return an integer less than, equal to, or greater than zero` warning.

This warning has now been mitigated by a slightly changed implementation, which will return an integer, as expected.

Additionally, as highlighted by the test expansion in the previous commit, the sort order for when the visibility of two taxonomies was considered "equal" was previously undefined.

This commit also fixes that and will now sort two taxonomies with the same visibility, case-sensitively by their public label using natural sort order.

The unit test has been adjusted to reflect the improve sorting.

Refs:
* https://wiki.php.net/rfc/stable_sorting
* https://www.php.net/manual/en/function.usort.php
* https://www.php.net/manual/en/function.strnatcmp.php