Check_Changes_Handler_Test: fix bug
This bug was exposed by PHP 8.1.

The `Check_Changes_Handler::check_changes_action_handler()` expects a global `$wp_version` variable to be available, but this variable was not created for the test, leading to a "Deprecated: version_compare(): Passing null to parameter #1 ($version1) of type string is deprecated" deprecation notices in PHP 8.1.

Fixed by improving the test.