CI: switch to GitHub Actions - step 2: add security check
As this repo does not have a committed `composer.lock` file, we can just require the `roave/security-advisories` repository to handle the security check.

The CS, Lint and Test workflows all run a `composer install`, which will fail if with that package in place if it would be attempted to install an insecure dependency.