Composer: allow composer installers 2.x
While the latest version of the `composer/installers` package is `2.2`, I'm explicitly setting the version constraint to a lenient one as this is a non-`dev` requirement, which comes into play when people use a [WP]Packagist based site setup.

If we set the requirement very strictly, it could conflict with other plugins with different constraints and it is not as if any of the 2.x releases have touched the WordPress installer, so setting it to `|| 2.0` should be fine.