Composer/Tests: update test dependencies
This updates the minimum version of the WP Test Utils dependency.

This also allows us to remove the work-around which was in place for Mockery not handling dynamic properties on mocked objects.
This was fixed in Mockery 1.6.0, which will be used when running the tests against PHP 8.2+.

Note: PHPUnit Polyfills 1.1.0 includes a new polyfill for the `assertObject[Not]HasProperty()` methods.
The assertions which these new assertions replace are not used in this test suite, so no test changes are needed.