Newsletter::newsletter_signup_form(): bug fix
The `wp_nonce_field()` function by default _displays_ the field.

This means that, as things were, the nonce field would be echo'd out when the function was called + would add a `1` to the `$html` string being created.
The resulting `$html` string would not contain the nonce field and the nonce field would not be between the `<form>` tags.

Fixed by setting the `$display` parameter to `false`.

Ref: https://developer.wordpress.org/reference/functions/wp_nonce_field/