Tests: consistency in terminology
The "integration" tests are being renamed to "WP (unit) tests".

This commit updates the references to these tests in the Composer scripts and the GH Actions workflows to follow the new terminology.

Note: the "old" `integration-test` and `integration-coverage` Composer scripts will still be available for the time being to facilitate the switch over and give the teams a little time to get used to the new script names.